读写分离

+++ Service/OrderService/src/main/java/com/mall/oms/entity/OrderInfo.java
@@ -38,7 +38,7 @@ public class OrderInfo implements Serializable {
     //收货信息
     private String orderUserName;
 
-    private Integer orderUserId;
+    private Long orderUserId;
 
     private String province;
